\name{chr2chrom}
\alias{chr2chrom}
\title{Converts a vector of Chr (numeric) values to Chrom (character) values.}
\usage{
chr2chrom(cs = NULL, chrvalues = NULL)
}
\arguments{
\item{cs}{A 'cs' object, either loaded from datasets in the current package, or generated through 'read_ucsc_cytobandideo'.}
\item{chrvalues}{A vector of Chr (numeric) values.}
}
\description{
Converts a vector of Chr (numeric) values to Chrom (character) values.
}
\author{
Bastien Job
}

