\name{chrLoad}
\alias{chrLoad}
\title{Loads chromosomal information of a species using its genome build name.}
\usage{
chrLoad(genome = "hg19")
}
\arguments{
\item{genome}{A genome build ID that will allow the loading of the corresponding object.}
}
\description{
Loads chromosomal information of a species using its genome build name. Currently available data are for genomes : hg17 to hg38, mm7 to mm10, and rn5 to rn6.
}
\author{
Bastien Job
}

