\name{chrom2chr}
\alias{chrom2chr}
\title{Converts a vector of Chrom (character) values to Chr (numeric) values.}
\usage{
chrom2chr(cs = NULL, chromvalues = NULL)
}
\arguments{
\item{cs}{A 'cs' object, either loaded from datasets in the current package, or generated through 'read_ucsc_cytobandideo'.}
\item{chromvalues}{A vector of Chrom (character) values.}
}
\description{
Converts a vector of Chrom (character) values to Chr (numeric) values.
}
\author{
Bastien Job
}

